Move movie-profile into stand-alone action reducer
