+++ src/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
-import logo from './logo.svg';
 import './App.css';
 
 import store from './init-store';
