+++ src/action/movie.js
@@ -6,12 +6,10 @@ import http from '../utils/http';
 export const REQUEST_STARTED = 'REQUEST_STARTED';
 export const REQUEST_FAIL = 'REQUEST_FAIL';
 export const BROWSE_REQUEST_SUCCESS = 'BROWSE_REQUEST_SUCCESS';
-export const BROWSE_DETAIL_REQUEST_SUCCESS = 'BROWSE_DETAIL_REQUEST_SUCCESS';
 
 const requestStarted = createAction(REQUEST_STARTED);
 const requestFail = createAction(REQUEST_FAIL);
 const browseRequestSuccess = createAction(BROWSE_REQUEST_SUCCESS);
-const browseDetailRequestSuccess = createAction(BROWSE_DETAIL_REQUEST_SUCCESS);
 
 type SearchQueriesType = {
   page: number,
@@ -82,20 +80,4 @@ export const searchMovies = (
   };
 
   dispatch(browseMovie(uri, queries));
-};
-
-// Async function to get movie details based on movie id
-export const getMovieDetails = (id: number) => async (dispatch: Function) => {
-  // Loading indicator for profile view
-  dispatch(requestStarted());
-
-  const uri = `/movie/${id}`;
-
-  try {
-    const browseResult = await http.GET(uri);
-
-    dispatch(browseDetailRequestSuccess(browseResult));
-  } catch (err) {
-    dispatch(requestFail(err));
-  }
 };
